fix master details vue between patient and symptoms
