+++ src/components/Quizz.vue
@@ -6,10 +6,10 @@
           <div class="panel">
             <div class="panel-header">
               <div class="panel-title column" align="right">
-                <h1
+                <h1 v-if="patient!=null"
                   class="title is-6 columns col-mx-auto default label"
                   align="right"
-                >Fiche de Symptômes</h1>
+                > {{patient.firstname + " "+patient.lastname + " "+ new Date().toISOString().slice(0,10)}}</h1>
               </div>
             </div>
             <div class="panel-nav">
@@ -49,23 +49,6 @@
                     >{{ response.text }}</div>
                   </template>
 
-                  <div v-if="quiz.questions[questionIndex].QUIZZ_TYPE=='FORMULAIRE'">
-                    <div
-                      v-for="(formulaire, index) in quiz.questions[questionIndex].formulaires"
-                      :key="index"
-                    >
-                      <label class="form-label" for="formulaire.label">{{formulaire.label}}</label>
-                      <input
-                        dir="rtl"
-                        class="form-input"
-                        :id="formulaire.label"
-                        type="text"
-                        :placeholder="formulaire.label"
-                        :value=" quiz.user "
-                      />
-                    </div>
-                  </div>
-
                   <div v-if="quiz.questions[questionIndex].QUIZZ_TYPE=='MULTIPLECHOICES'">
                     <div
                       v-for="(response, index) in quiz.questions[questionIndex].responses"
@@ -95,15 +78,11 @@
                 </span>
 
                 <!--resultTitleBlock-->
-                <h2 class="title">Best Score</h2>
+                <h4 class="title" align="left">Votre questionnaire et terminé vous pouvez envoyer le formaulaire maintenant</h4>
                 <!--<p class="subtitle">
 					Total score: {{ score() }} / {{ quiz.questions.length }}
                 </p>-->
                 <br />
-                <a class="button" @click="restart()">
-                  restart
-                  <i class="fa fa-refresh"></i>
-                </a>
                 <!--/resultTitleBlock-->
               </div>
             </div>
@@ -123,7 +102,7 @@
                   <!-- centered logo or brand    || userResponses[questionIndex]==null -->
                 </section>
                 <section class="navbar-section">
-                  <button
+                  <button v-if="questionIndex<quiz.questions.length"
                     :class="(userResponses[questionIndex]==null)?'':'is-active'"
                     v-on:click="next();"
                     :disabled="questionIndex>=quiz.questions.length  "
@@ -132,12 +111,18 @@
                     Suivant
                     <i class="icon icon-arrow-right"></i>
                   </button>
+
+                   <button v-if="questionIndex>=quiz.questions.length"
+                    v-on:click="save();"
+                    class="btn btn-success"
+                  >
+                    Envoyer maintenant
+                    <i class="icon icon-mail"></i>
+                  </button>
+
                 </section>
               </header>
               <!--/quizFooter-->
-              <div class="form-group">
-                <button @click="exit()" class="btn btn-link">Quitter</button>
-              </div>
             </div>
           </div>
         </div>
@@ -148,8 +133,15 @@
 
 <script>
 import Vue from "vue";
+import questions  from '../assets/questions';
+var quizz = questions;
+let userResponseSkelaton = Array(quizz.questions.length).fill(null);
 export default {
   name: "Quizz",
+  props:
+  {
+    patient : Object
+  },
   data() {
     return {
       message: "hello",
@@ -262,9 +254,6 @@ export default {
         }
       }
     },
-    exit: function() {
-      this.$router.push("patients");
-    },
     // Return "true" count in userResponses
     score: function() {
       var score = 0;
@@ -283,240 +272,7 @@ export default {
     }
   }
 };
-var quizz = {
-    user: "",
-    questions: [
-      {
-        ID: "ASPHEXIE",
-        QUIZZ_TYPE: "ONECHOICE",
-        text: " هل تشكو من عدم القدرة على الحديث كما في السابق ؟ ",
-        responses: [
-          { text: " نعم ", correct: true, code: "YES" },
-          { text: " لا  ", code: "NO" }
-        ]
-      },
-      {
-        ID: "ASPHEXIE_CAUSE",
-        QUIZZ_TYPE: "MULTIPLECHOICES",
-        text: " لماذا ؟ ",
-        responses: [
-          { text: " احس بضيق في التنفس ", code: "1", selected: false },
-          { text: "احس بتعب شديد ", code: "2", selected: false },
-          { text: "صداع شديد يمنعني من الحديث", code: "3", selected: false },
-          { text: " أسباب أخرى ", code: "4", selected: false }
-        ],
-        isvisible: function(id, code) {
-          return id == "ASPHEXIE" && code == "YES" ? true : false;
-        }
-      },
-      {
-        ID: "USER",
-        text: "من يقوم بتعمير هذه الاستمارة ؟ ",
-        QUIZZ_TYPE: "ONECHOICE",
-        responses: [
-          { text: " المريض", correct: true, code: "YES" },
-          { text: "شخص آخر", code: "NO" }
-        ]
-      },
-      {
-        ID: "USERCAUSE",
-        QUIZZ_TYPE: "ONECHOICE",
-        text: "لماذا يقوم شخص آخر بملأ الإستمارة ؟ ",
-        responses: [
-          { text: " المريض لا يعرف التعامل مع التطبيقة ", correct: true },
-          {
-            text:
-              "حالة المريض  الصحية لا تسمح جراء فقدانه الوعي أو ضيق تنفس حاد"
-          }
-        ],
-        isvisible: function(id, code) {
-          return id == "USER" && code == "NO" ? true : false;
-        }
-      },
-      {
-        ID: "FORMULAIRE",
-        QUIZZ_TYPE: "FORMULAIRE",
-        text: " : الرجاء قم بإدخال المعطيات االتالية",
-        responses: [{ text: "Oui", correct: true }, { text: "Non" }],
-        formulaires: [
-          { label: "Nom", type: "text", reponse: "" },
-          { label: "Prenom", type: "text", reponse: "" }
-        ]
-      },
-      {
-        ID: "DIABETE",
-        QUIZZ_TYPE: "ONECHOICE",
-        text: "هل تعاني من مرض السكري ؟",
-        responses: [
-          { text: "نعم", correct: true, code: "YES" },
-          { text: " لا ", code: "NO" }
-        ]
-      },
-      {
-        text: " هل تعاني من ضغط الدم ؟",
-        QUIZZ_TYPE: "ONECHOICE",
-        responses: [
-          { text: "نعم", correct: true, code: "YES" },
-          { text: " لا ", code: "NO" }
-        ]
-      },
-      {
-        ID: "CARDIO",
-        QUIZZ_TYPE: "ONECHOICE",
-        text: "  هل تعاني من  أمراض القلب ؟ ",
-        responses: [
-          { text: "نعم", correct: true, code: "YES" },
-          { text: " لا ", code: "NO" }
-        ]
-      },
-      {
-        ID: "ASME",
-        QUIZZ_TYPE: "ONECHOICE",
-        text: " هل تعاني من ضيق تنفس مزمن ؟ ",
-        responses: [
-          { text: "نعم", correct: true, code: "YES" },
-          { text: " لا ", code: "NO" }
-        ]
-      },
-      {
-        ID: "RENALE",
-        QUIZZ_TYPE: "ONECHOICE",
-        text: "هل تعاني من  أمراض الكلى ؟ ",
-        responses: [
-          { text: "نعم", correct: true, code: "YES" },
-          { text: " لا ", code: "NO" }
-        ]
-      },
-      {
-        ID: "CANCER",
-        QUIZZ_TYPE: "ONECHOICE",
-        text: "هل أصبت بمرض السرطان أو تتبع   علاجا  بشأنه ؟ ",
-        responses: [
-          { text: "نعم", correct: true, code: "YES" },
-          { text: " لا ", code: "NO" }
-        ]
-      },
-      {
-        ID: "IMMUNITY",
-        QUIZZ_TYPE: "ONECHOICE",
-        text: " هل تعاني من أحد أمراض ضعف المانعة ؟ ",
-        responses: [
-          { text: "نعم", correct: true, code: "YES" },
-          { text: " لا ", code: "NO" }
-        ]
-      },
-      {
-        ID: "VOYAGE",
-        QUIZZ_TYPE: "ONECHOICE",
-        text: "هل سافرت خارج البلاد خلال الشهر الفارط ؟ ",
-        responses: [
-          { text: " نعم ", correct: true, code: "YES" },
-          { text: " لا  ", code: "NO" }
-        ]
-      },
-      {
-        ID: "EXPOSITION_VOYAGE",
-        QUIZZ_TYPE: "ONECHOICE",
-        text: " هل كان لك اتصال مباشر بشخص سافر خارج البلاد الشهر الفارط ؟ ",
-        responses: [
-          { text: " نعم ", correct: true, code: "YES" },
-          { text: " لا  ", code: "NO" }
-        ],
-        isvisible: function(id, code) {
-          return id == "VOYAGE" && code == "NO" ? true : false;
-        }
-      },
-      {
-        ID: "EXPOSITION_HABITAT",
-        QUIZZ_TYPE: "ONECHOICE",
-        text: "هل يسكن معك في نفس المنزل ؟",
-        responses: [
-          { text: " نعم ", correct: true, code: "YES" },
-          { text: " لا  ", code: "NO" }
-        ],
-        isvisible: function(id, code) {
-          return id == "EXPOSITION_VOYAGE" && code == "YES" ? true : false;
-        }
-      },
-      {
-        ID: "EXPOSITION_ANALYSE",
-        QUIZZ_TYPE: "ONECHOICE",
-        text: " هل خضعت أنت او الشخص المذكور لتحليل الاصابة بفيروس كورونا ؟",
-        responses: [
-          { text: " نعم ", correct: true, code: "YES" },
-          { text: " لا  ", code: "NO" }
-        ],
-        isvisible: function(id, code) {
-          return id == "EXPOSITION_VOYAGE" && code == "YES" ? true : false;
-        }
-      },
-      {
-        ID: "EXPOSITION_RESULTAT_ANALYSE",
-        QUIZZ_TYPE: "ONECHOICE",
-        text: "ماهي نتيجة التحليل ؟",
-        responses: [
-          { text: "  مصاب ", correct: true, code: "YES" },
-          { text: "غير مصاب", code: "NO" }
-        ],
-        isvisible: function(id, code) {
-          return id == "EXPOSITION_ANALYSE" && code == "YES" ? true : false;
-        }
-      },
 
-      {
-        ID: "FIEVRE",
-        QUIZZ_TYPE: "ONECHOICE",
-        text: " هل تشكو من ارتفاع في درجة حرارتك (حمى) ؟ ",
-        responses: [
-          { text: " نعم ", correct: true, code: "YES" },
-          { text: " لا  ", code: "NO" }
-        ]
-      },
-      {
-        ID: "MAUX_GORGE",
-        QUIZZ_TYPE: "ONECHOICE",
-        text: " هل تشعر بألم حاد في الحنجرة ؟ ",
-        responses: [
-          { text: " نعم ", correct: true, code: "YES" },
-          { text: " لا  ", code: "NO" }
-        ]
-      },
-      {
-        ID: "TOUX",
-        QUIZZ_TYPE: "ONECHOICE",
-        text:
-          " هل لديك سعال او ان درجة السعال التي تشكو منها في العادة ازدادت حدة ؟ ",
-        responses: [
-          { text: " نعم ", correct: true, code: "YES" },
-          { text: " لا  " }
-        ]
-      },
-      {
-        ID: "ABILYTY",
-        QUIZZ_TYPE: "ONECHOICE",
-        text: "هل تستطيع الحركة و القيام بحاجياتك دون مساعدة ؟",
-        responses: [
-          { text: " نعم ", correct: true, code: "YES" },
-          { text: " لا  ", code: "NO" }
-        ]
-      },
-      {
-        ID: "ABILYTY_DURATION",
-        QUIZZ_TYPE: "ONECHOICE",
-        text: " منذ متى ؟",
-        responses: [
-          { text: " ايام ", correct: true },
-          { text: "اسابيع ", correct: true },
-          { text: "اشهر " },
-          { text: "سنوات " }
-        ],
-        isvisible: function(id, code) {
-          return id == "ABILYTY" && code == "NO" ? true : false;
-        }
-      }
-    ]
-  },
-  userResponseSkelaton = Array(quizz.questions.length).fill(null);
 </script>
 
 <style scoped>
